Move config.yml to config subdirectory
Allows to use a folder as docker volume (eg. NFS share)
