+++ Dockerfile
@@ -41,7 +41,7 @@ RUN apk add --no-cache tzdata
 # Copy built application from build phase
 COPY --from=BUILD_IMAGE /app ./
 # Ensure only one version of conf.yml exists
-RUN rm dist/conf.yml
+RUN rm dist/config/conf.yml
 
 # Finally, run start command to serve up the built application
 CMD [ "yarn", "start" ]
