+++ services/config-validator.js
@@ -99,7 +99,7 @@ const printFileReadError = (e) => {
 };
 
 try { // Try to open and parse the YAML file
-  const config = yaml.load(fs.readFileSync('./public/conf.yml', 'utf8'));
+  const config = yaml.load(fs.readFileSync('./public/config/conf.yml', 'utf8'));
   validate(config);
 } catch (e) { // Something went very wrong...
   setIsValidVariable(false);
